% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datSumDWM.R
\name{datSumDWM}
\alias{datSumDWM}
\title{Data - Aggregates numeric cover data to the plot or condition-level.}
\usage{
datSumDWM(
  dwm,
  dwmsumvar,
  plt = NULL,
  cond = NULL,
  datsource = "obj",
  dbconn = NULL,
  dsn = NULL,
  bycond = FALSE,
  bydomainlst = NULL,
  getadjplot = FALSE,
  domclassify = NULL,
  pltidsWITHqry = NULL,
  pltidsid = NULL,
  pcwhereqry = NULL,
  savedata = FALSE,
  tabIDs = tableIDs(),
  datSum_opts = datSum_options(),
  database_opts = NULL,
  savedata_opts = NULL
)
}
\arguments{
\item{dwm}{Dataframe or comma-delimited file (*.csv). The table with
down woody material (dwm) data.}

\item{dwmsumvar}{String. Name of estimation variable in dwm (e.g., CWD_DRYBIO_UNDAJ)}

\item{plt}{Dataframe, comma-delimited file (*.csv), or shapefile (*.shp).
Plot-level table to join the aggregated tree data to, if bycond=FALSE. This
table is optional.}

\item{cond}{Dataframe or comma-delimited file (*.csv). Condition-level table
to join the aggregated tree data to, if bycond=TRUE. This table also may be
used for condition proportion or strata variables used if adjcond or
adjstrata = TRUE (See details below).  This table is optional.}

\item{datsource}{String. Source of data ('obj', 'csv', 'sqlite', 'gdb').}

\item{dbconn}{Open database connection.}

\item{dsn}{String. If datsource='sqlite', the name of SQLite database
(*.sqlite).}

\item{bycond}{Logical. If TRUE, the data are aggregated to the condition
level (by: cuniqueid, condid). If FALSE, the data are aggregated to the plot
level (by: puniqueid).}

\item{bydomainlst}{String (vector). Categorical domain variables for
summing tree data by (e.g., SPCD). Variables must be in tree table or
plt/cond table if tables are provided.}

\item{getadjplot}{Logical. If TRUE, and adj='plot', adjfactors are
calculated for nonsampled conditions at plot-level.}

\item{domclassify}{List. List for classifying domain variables in bydomainlst
(e.g., DIA = c(10,20,30)).}

\item{pltidsWITHqry}{SQL query. A query identifying plots to sum (e.g.,
'WITH pltids AS (SELECT cn AS PLT_CN FROM plot WHERE statecd=49 and INVYR=2018)')}

\item{pltidsid}{Sting. Name of unique identifier in pltidsWITHqry.}

\item{pcwhereqry}{String. Plot/Condition filter if plot and/or cond table is
included.}

\item{savedata}{Logical. If TRUE, saves data to outfolder.}

\item{tabIDs}{List of unique IDs corresponding to the tables. See
See help(tableIDs) for a list of options.}

\item{datSum_opts}{List. Options for summarizing tree data, such as TPA,
rounding, and adjusting TPA. See help(datSum_options()) for a list of
options.}

\item{database_opts}{List. Options for database, such as schema and
password. See help(database_options()) for a list of options.}

\item{savedata_opts}{List. See help(savedata_options()) for a list}
}
\value{
A list of the following items: \item{dwmdat}{ Data frame. Plot or
condition-level table with aggregated tree attributes. } \item{sumvars}{
String vector. Name(s) of the output aggregated tree attributes. }

If savedata=TRUE\cr - treedat will be saved to the outfolder. \cr - a text
file of input parameters is saved to outfolder
('outfn'_parameters_'date'.txt).
}
\description{
Aggregates percent cover data (e.g., COVER_PCT) to plot or condition,
including options for filtering.
}
\details{
If you want to adjust plot-level or subplot-level information by condition
proportions (adjplot), you need to include CONDID & CONDPROP_UNADJ in cond
or tree table and COND_STATUS_CD. \cr
}
\note{
If a dat table is provided, the aggregated data will be merged to
table and NULL values will be output as 0.
}
\author{
Tracey S. Frescino
}
\keyword{data}
