% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datSumTreeGRM.R
\name{datSumTreeGRM}
\alias{datSumTreeGRM}
\title{Data - Aggregates numeric tree data to the plot or condition-level.}
\usage{
datSumTreeGRM(
  grmtables = list(plt = "plot", cond = "cond", tree = "tree", tree_grm_component =
    "tree_grm_component", tree_grm_begin = "tree_grm_begin", tree_grm_midpt =
    "tree_grm_midpt", beginend = "beginend"),
  datsource = "obj",
  dbconn = NULL,
  dsn = NULL,
  landarea,
  tstatus,
  grmtyp,
  tsumvar,
  tsumvarnm = "ESTIMATED_VALUE",
  tdomvar = NULL,
  tdomvarlst = NULL,
  tdomvar2 = NULL,
  tdomvar2lst = NULL,
  bycond = FALSE,
  bydomainlst = NULL,
  grm_typ_cd = 2,
  pivot = TRUE,
  domclassify = NULL,
  pltidsWITHqry = NULL,
  pltidsid = NULL,
  pltcondx = NULL,
  pltidsadj = NULL,
  savedata = FALSE,
  tabIDs = tableIDs(),
  datSum_opts = datSum_options(),
  database_opts = NULL,
  savedata_opts = NULL
)
}
\arguments{
\item{grmtables}{List of tables used for summarizing grm data.}

\item{datsource}{String. Source of data ('obj', 'csv', 'sqlite', 'gdb').}

\item{dbconn}{Open database connection.}

\item{dsn}{String. If datsource='sqlite', the name of SQLite database
(*.sqlite).}

\item{landarea}{String. The sample area filter for estimates ("ALL",
"FOREST", "TIMBERLAND").  If landarea=FOREST, filtered to COND_STATUS_CD =
1; If landarea=TIMBERLAND, filtered to SITECLCD in(1:6) and RESERVCD = 0.}

\item{tstatus}{String. The status type of GRM to summarize ('Sawtimber',
'Growing-stock', 'All live').}

\item{grmtyp}{String. Type of GRM to summarize ('GROW', 'MORT', 'REMV')}

\item{tsumvar}{String. Tree-level variable to summarize (e.g., "VOLCFNET").}

\item{tsumvarnm}{String. Name of summarized output column. The default
is 'ESTIMATED_VALUE'.}

\item{tdomvar}{String. The tree domain (tdom) variable used to aggregate by
(e.g., "SPCD", "SPGRPCD").}

\item{tdomvarlst}{String (vector). List of specific tree domains of tdomvar
to aggregate (e.g., c(108, 202)). If NULL, all domains of tdomvar are used.}

\item{tdomvar2}{String. A second tree domain variable to use to aggregate by
(e.g. "DIACL").  The variables, tdomvar and tdomvar2 will be concatenated
before summed.}

\item{tdomvar2lst}{String (vector). List of specific tree domains of
tdomvar2 to aggregate.  If NULL, all domains of tdomvar2 are used.}

\item{bycond}{Logical. If TRUE, the data are aggregated to the condition
level (by: cuniqueid, condid). If FALSE, the data are aggregated to the plot
level (by: puniqueid). If bysubp = TRUE and bycond = TRUE, data are
aggregated by subplotid, subpid, condid.}

\item{bydomainlst}{String (vector). Categorical domain variables for
summing tree data by (e.g., SPCD). Variables must be in tree table or
plt/cond table if tables are provided.}

\item{grm_typ_cd}{Integer Type of annual GRM code (1:Current annual; 2:A2A)}

\item{pivot}{Logical. If TRUE, tdomvar data are transposed (pivoted) to
separate columns.}

\item{domclassify}{List. List for classifying domain variables in bydomainlst
(e.g., DIA = c(10,20,30)).}

\item{pltidsWITHqry}{SQL query. A query identifying plots to sum (e.g.,
'WITH pltids AS (SELECT cn AS PLT_CN FROM plot WHERE statecd=49 and INVYR=2018)')}

\item{pltidsid}{Sting. Name of unique identifier in pltidsWITHqry.}

\item{savedata}{Logical. If TRUE, saves data to outfolder.}

\item{tabIDs}{List of unique IDs corresponding to the tables. See
See help(tableIDs) for a list of options.}

\item{datSum_opts}{List. Options for summarizing tree data, such as TPA,
rounding, and adjusting TPA. See help(datSum_options()) for a list of
options.}

\item{database_opts}{List. Options for database, such as schema and
password. See help(database_options()) for a list of options.}

\item{savedata_opts}{List. See help(savedata_options()) for a list}
}
\value{
A list of the following items: \item{treedat}{ Data frame. Plot or
condition-level table with aggregated tree attributes. } \item{sumvars}{
String vector. Name(s) of the output aggregated tree attributes. }
}
\description{
Aggregates numeric tree-level data (e.g., VOLCFNET) to plot or condition,
including options for filtering tree data or extrapolating to acre by
multiplying by TPA.
}
\details{
For adjcond (bycond=FALSE): \cr If you want to summarize trees-per-acre
information aggregated to plot or condition level, you need to include a TPA
variable in tree table. \cr For tsumvars = GROWCFGS, GROWBFSL, GROWCFAL,
FGROWCFGS, FGROWBFSL, or FGROWCFAL, you must have TPAGROW_UNADJ \cr For
tsumvars = MORTCFGS, MORTBFSL, MORTCFAL, FMORTCFGS, FMORTBFSL, or FMORTCFAL,
you must have TPAMORT_UNADJ \cr For tsumvars = REMVCFGS, REMVBFSL, REMVCFAL,
FREMVCFGS, FREMVBFSL, or FREMVCFAL, you must have TPAREMV_UNADJ \cr

If you want to adjust plot-level or subplot-level information by condition
proportions (adjplot), you need to include CONDID & CONDPROP_UNADJ in cond
or tree table and COND_STATUS_CD. \cr
}
