% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBgetXY.R
\name{DBgetXY}
\alias{DBgetXY}
\title{Database - Extracts actual plot coordinates from the FIA Oracle database.}
\usage{
DBgetXY(
  states = NULL,
  RS = NULL,
  datsource = "datamart",
  dsn = NULL,
  xy = "PLOT",
  xy_opts = xy_options(xy.uniqueid = "CN", xvar = "LON", yvar = "LAT"),
  dbTabs = dbTables(),
  eval = "FIA",
  eval_opts = eval_options(),
  pjoinid = "CN",
  invtype = "ANNUAL",
  intensity1 = FALSE,
  issp = FALSE,
  addINTENSITY = TRUE,
  addINVYR = TRUE,
  addMEASYEAR = FALSE,
  returndata = TRUE,
  savedata = FALSE,
  exportsp = FALSE,
  savedata_opts = NULL,
  dbconn = NULL,
  dbconnopen = FALSE
)
}
\arguments{
\item{states}{String or numeric vector. Name (e.g., 'Arizona','New Mexico')
or code (e.g., 4, 35) of state(s) for evalid. If all states in one or more
FIA Research Station is desired, set states=NULL and use RS argument to
define RS.}

\item{RS}{String vector. Name of research station(s) to get public XY
coordinates for ('RMRS','SRS','NCRS','NERS','PNWRS'). Do not use if states 
is populated. See FIESTA::ref_statecd for reference to RS and states.}

\item{datsource}{Source of data ('datamart', 'sqlite').}

\item{dsn}{If datsource='sqlite', the file name (data source name) of
the sqlite database (*.sqlite).}

\item{xy}{sf R object or String. Table with xy coordinates. Can be a spatial
polygon object, data frame, full pathname to a shapefile, or name of a layer
within a database.}

\item{xy_opts}{List of xy data options to specify if xy is NOT NULL. 
See xy_options (e.g., xy_opts = list(xvar='LON', yvar='LAT').}

\item{dbTabs}{List of database tables the user would like returned.
See help(dbTables) for a list of options.}

\item{eval}{String. Type of evaluation time frame for data extraction 
('FIA', 'custom'). See eval_opts for more further options.}

\item{eval_opts}{List of evaluation options for 'FIA' or 'custom'
evaluations to determine the set of data returned. See help(eval_options)
for a list of options.}

\item{pjoinid}{String. Variable in plt to join to XY data. Not necessary to
be unique. If using most current XY coordinates, use identifier for a plot
(e.g., PLOT_ID).}

\item{invtype}{String. Type of FIA inventory to extract ('PERIODIC',
'ANNUAL').  Only one inventory type (PERIODIC/ANNUAL) at a time.}

\item{intensity1}{Logical. If TRUE, includes only XY coordinates where 
INTENSITY = 1 (FIA base grid).}

\item{issp}{Logical. If TRUE, returns spatial XY data as a list object with
query.}

\item{addINTENSITY}{Logical. If TRUE, appends INTENSITY variable to dataframe.}

\item{addINVYR}{Logical. If TRUE, appends INVYR variable to dataframe.}

\item{addMEASYEAR}{Logical. If TRUE, appends INTENSITY variable to dataframe.}

\item{returndata}{Logical. If TRUE, returns XY data as a list object with
query.}

\item{savedata}{Logical. If TRUE, saves XY data to outfolder as comma-delimited
file (*.csv).}

\item{exportsp}{Logical. If TRUE, exports data as spatial.}

\item{savedata_opts}{List. See help(savedata_options()) for a list
of options. Only used when savedata = TRUE.}

\item{dbconn}{Open database connection.}

\item{dbconnopen}{Logical. If TRUE, the dbconn connection is not closed.}
}
\value{
if returndata=TRUE, a list of the following objects: 
\item{xy*_PUBLIC}{ Data frame. XY data from FIA's public database. If 
measCur=TRUE, named xyCur_PUBLIC, else named xy_PUBLIC. The data frame 
has 10 columns ('PLT_CN', 'LON_PUBLIC', 'LAT_PUBLIC', 'STATECD', 'UNITCD',
'COUNTYCD', 'PLOT', 'INTENSITY', 'PLOT_ID' (ID+STATECD+UNTCD+COUNTYCD+PLOT), 
'COUNTYFIPS'. If issp=TRUE, returns an sf object. }
\item{xyqry}{ String. Query to extract coordinates. }
\item{xvar}{ String. Name of X variable in xy*_PUBLIC. }
\item{yvar}{ String. Name of Y variable in xy*_PUBLIC. }

If savedata=TRUE, outputs the xy*_PUBLIC as out_fmt to outfolder. 
If exportsp=TRUE, the output xy data are saved as spatial layer.
}
\description{
Extracts actual plot coodinates for the most current measurement of each
plot in the database.
}
\note{
If no parameters are included, the user is prompted for input. If partial
parameters, the default parameter values are used for those not specified.
}
\examples{
\dontrun{
# Most current evaluation and shapefile with public coordinates
COxylst <- DBgetXY(states = "Colorado",
                   eval_opts=eval_options(measCur = TRUE))
names(COxylst)

head(COxylst$xyCur_ACTUAL)
COxylst$xyqry
}
}
\author{
Tracey S. Frescino
}
\keyword{data}
