% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTables.R
\name{dbTables}
\alias{dbTables}
\title{List of population tables.}
\usage{
dbTables(
  plot_layer = "PLOT",
  cond_layer = "COND",
  tree_layer = "TREE",
  seed_layer = "SEEDLING",
  plotgeom_layer = "PLOTGEOM",
  vsubpspp_layer = "P2VEG_SUBPLOT_SPP",
  vsubpstr_layer = "P2VEG_SUBP_STRUCTURE",
  invsubp_layer = "INVASIVE_SUBPLOT_SPP",
  subplot_layer = "SUBPLOT",
  subpc_layer = "SUBP_COND",
  dwm_layer = "COND_DWM_CALC",
  sccm_layer = "SUBP_COND_CHNG_MTRX",
  grm_layer = "TREE_GRM_COMPONENT",
  survey_layer = "SURVEY",
  ppsa_layer = "POP_PLOT_STRATUM_ASSGN",
  ...
)
}
\arguments{
\item{...}{For extendibility.}

\item{plot_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Plot-level data.}

\item{cond_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Condition-level data.}

\item{seed_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Seedling data.}

\item{plotgeom_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Plot-level GIS extracted data.}

\item{vsubpspp_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Understory vegetation species data.}

\item{vsubpstr_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Understory vegetation structure data.}

\item{invsubp_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Understory vegetation invasives data.}

\item{subplot.}{R object, comma-delimited file(*.csv), or name of
layer in database. Subplot-level data.}

\item{subpc_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Subplot condition-level data.}

\item{dwm_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Down wood material data.}

\item{sccm_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Subplot-level change matrix data.}

\item{grm_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Tree growth, removal, mortality data.}

\item{survey_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Population survey data.}

\item{ppsa_layer.}{R object, comma-delimited file(*.csv), or name of
layer in database. Population plot stratum assignment data.}

\item{other_layers.}{R object, comma-delimited file(*.csv), or name of
layer in database. Other layers to extract from database.}
}
\value{
A list of user-supplied parameters and parameter values for strata.
}
\description{
Returns a list of user-supplied parameters and parameter values for data 
tables to be supplied to *DB functions.
}
\details{
If no parameters, an empty list is returned.
}
\examples{
dbTables(sccm_layer)
}
\author{
Tracey S. Frescino
}
\keyword{list}
