\name{DBtestPostgreSQL}
\alias{DBtestPostgreSQL}
\title{
	Database - Test a SQLite database table. 
}
\description{
	Checks a SQLite database.
}
\usage{
	DBtestPostgreSQL(dbname, dbconnopen = TRUE)
}
 
\arguments{
  \item{dbname}{ String. Name of PostgreSQL database. }
  \item{dbconnopen}{ Logical. If TRUE, the dbconn connection is not closed. }
}
\author{
	Tracey S. Frescino
}


\keyword{ data }
