% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{DEFAULT_NODATA}
\alias{DEFAULT_NODATA}
\title{Reference tables - NODATA Values.}
\format{
An object of class \code{list} of length 7.
}
\source{
gdal values.
}
\usage{
DEFAULT_NODATA
}
\description{
Reference tables - NODATA Values.
}
\keyword{datasets}
