% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{GDT_NAMES}
\alias{GDT_NAMES}
\title{Reference tables - gdal data types.}
\format{
An object of class \code{character} of length 12.
}
\source{
gdal values.
}
\usage{
GDT_NAMES
}
\description{
Reference tables - gdal data types.
}
\keyword{datasets}
