\name{FIESTA-internal}
\alias{addCI}
\alias{addcommas}
\alias{DBgetvars}
\alias{change10char}
\alias{check.logic}
\alias{check.data}
\alias{check.dataPB}
\alias{check.nostrata}
\alias{check.auxiliary}
\alias{check.strata}
\alias{check.rowcol}
\alias{check.rowcolPB}
\alias{check.titles}
\alias{check.tree}
\alias{check.PROP}
\alias{check.unitarea}
\alias{check.matchclass}
\alias{check.matchval}
\alias{check.numeric}
\alias{check.pltcnt}
\alias{check.extents}
\alias{fileexistsnm}
\alias{est.outtabs}
\alias{getarea}
\alias{getadjfactorGB}
\alias{getadjfactorPLOT}
\alias{getcrossord}
\alias{GBest.pbar}
\alias{getfunnm}
\alias{getlt10char}
\alias{getrastlst}
\alias{getRefobject}
\alias{getrhat}
\alias{getShapedat}
\alias{getstratacnt}
\alias{groupEstunit}
\alias{groupStrata}
\alias{groupUnits}
\alias{makebrick}
\alias{pastevars}
\alias{pcheck.DBtabs}
\alias{pcheck.landarea}
\alias{pcheck.logical}
\alias{pcheck.outfolder}
\alias{pcheck.rowcol}
\alias{pcheck.states}
\alias{pcheck.nostrata}
\alias{pcheck.strata}
\alias{pcheck.table}
\alias{pcheck.unitacres}
\alias{pcheck.varchar}
\alias{pcheck.xlsx}
\alias{projfix}
\alias{save1tab}
\alias{save2tabs}
\alias{selectvars}
\alias{write2csv}
\alias{wtcalc}
\alias{wtcalc.unit}
\alias{bp_wrap.it}
\alias{bp_width.jpg}

\alias{removecols}
\alias{DT_NAto0}
\alias{changeNULL}
\alias{getdups}
\alias{getlistparam}
\alias{getnm}
\alias{checknm}
\alias{check.namedlist}
\alias{capfirst}
\alias{DBgetfn}
\alias{clip.othertables}

\alias{PBest.pbar}
\alias{PBest.pbarRatio}
\alias{PBgetest}
\alias{getpltdom.prop}
\alias{getgainloss}
\alias{transpose2row}
\alias{transpose2col}

\alias{add0unit}
\alias{addrowgrp}
\alias{allin1f}
\alias{crosstabx}
\alias{crossxtab}
\alias{getdomain}
\alias{getestvar}

\alias{wrapSE}
\alias{tabgrp}
\alias{setCells}
\alias{write2xlsx}

\alias{getRaster}
\alias{pcheck.spatial}
\alias{build.prj4str}
\alias{trunc10shp}
\alias{getEPSG}
\alias{check.extents} 
\alias{CRScompare}
\alias{getprjatt}
\alias{areacalc.poly}
\alias{areacalc.pixel}
\alias{aspect_transform}
\alias{spPlotRastcl}
\alias{writeESRIprj}

\alias{getOffset}
\alias{getGDALDataTypeName}
\alias{basename.NoExt}
\alias{Mode}
\alias{getPixelValue}
\alias{extractPtsFromRaster}
\alias{extractPtsFromRasterList}
\alias{rasterFromRaster}
\alias{RasterizePolygon}
\alias{rasterizePolygons}
\alias{clipRaster}
\alias{rasterCombine}
\alias{zonalMean}
\alias{zonalFreq}
\alias{zonalMajority}
\alias{zonalMinority}
\alias{zonalVariety}
\alias{CmbTable} 
\alias{eastness}
\alias{northness}
\alias{RunningStats}
\alias{getGDALformat}
\alias{ptCsvToVRT}
\alias{rasterToVRT}
\alias{readOGRfromSQL}
\alias{zonalStats}

\alias{MAest.unit}
\alias{MAest.dom}
\alias{MAest}
\alias{MAest.ps}
\alias{MAest.greg}
\alias{MAest.ht}
\alias{SAest.large}
\alias{SAest.dom}
\alias{SAest}



\title{ Internal FIESTA Functions }
\description{
  Internal Subfunctions
}

\details{
These are not to be called by the user.  
}

\author{ Tracey S. Frescino }

\keyword{internal}

