% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{kindcd3old}
\alias{kindcd3old}
\title{Reference table - List of RMRS plots that have fallen out of inventory
because they were not found or they were in the wrong place.}
\format{
An object of class \code{data.frame} with 38 rows and 8 columns.
}
\source{
FIA query.
SELECT bp.STATECD, bp.COUNTYCD, bp.PLOT_FIADB NEW_PLOT, bp.START_DATE NEW_START_DATE,
      	bp_old.COUNTYCD OLD_COUNTYCD, bp_old.PLOT_FIADB OLD_PLOT, 
bp_old.END_DATE OLD_END_DATE, p.CN
FROM fs_nims_rmrs.NIMS_BASE_PLOT bp
JOIN fs_nims_rmrs.NIMS_BASE_PLOT bp_old on (bp.PREV_NBP_CN=bp_old.CN)
JOIN fs_nims_rmrs.NIMS_PLOT_RMRS_VW p on(p.NBP_CN=bp_old.CN)
WHERE p.KINDCD = 1
ORDER BY bp.STATECD, bp.COUNTYCD, bp_old.PLOT_FIADB"
}
\usage{
kindcd3old
}
\description{
Reference table - List of RMRS plots that have fallen out of inventory
because they were not found or they were in the wrong place.
}
\keyword{datasets}
