% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modGBdomdat.R
\name{modGBdomdat}
\alias{modGBdomdat}
\title{Generates estimates from domain-level data.}
\usage{
modGBdomdat(
  domdat,
  uniqueid = NULL,
  estvar,
  rowvar = NULL,
  colvar = NULL,
  stratdat = NULL,
  savedata = FALSE,
  table_opts = NULL,
  title_opts = NULL,
  savedata_opts = NULL,
  ...
)
}
\arguments{
\item{domdat}{List. domain-level data for estimation}

\item{uniqueid}{string. name of unique identifier of plots in domdat (e.g., plt_cn)}

\item{estvar}{string. name of variable in domdat to estimate.}

\item{rowvar}{string. name of rowvar domain in domdat to estimate by.}

\item{colvar}{string. name of colvar domain in domdat to estimate by.}

\item{stratdat}{list. returned list from spGetStrata. If NULL, no strata
used for estimation (i.e., simple random sample)}

\item{savedata}{Logical. If TRUE, saves table(s) to outfolder.}

\item{table_opts}{List. See help(table_options()) for a list of
options.}

\item{savedata_opts}{List. See help(savedata_options()) for a list
of options. Only used when savedata = TRUE.}

\item{...}{Parameters for modGBpop() if GBpopdat is NULL.}
}
\value{
A list with estimates with percent sampling error for rowvar (and
colvar).  If sumunits=TRUE or unitvar=NULL and colvar=NULL, one data frame
is returned.  Otherwise, a list object is returned with the following
information.  If savedata=TRUE, all data frames are written to outfolder.
}
\description{
Generates estimates from domain-level data.
}
\keyword{data}
