% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popFilters.R
\name{popFilters}
\alias{popFilters}
\title{Population data filters.}
\usage{
popFilters(
  evalid = NULL,
  invyrs = NULL,
  intensity = NULL,
  measCur = NULL,
  measEndyr = NULL,
  measEndyr.filter = NULL,
  ACI = FALSE,
  ...
)
}
\arguments{
\item{evalid}{}

\item{invyrs}{}

\item{intensity}{}

\item{measCur}{}

\item{measEndyr}{}

\item{measEndyr.filter}{}

\item{ACI}{}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for population
data filters.
}
\description{
Returns a list of user-supplied parameters and parameter values for population
data filters.
}
\details{
If no parameters, an empty list is returned.
}
\examples{

popFilters(ACI = TRUE)

}
\author{
Grayson W. White
}
\keyword{filters}
