% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popTableIDs.R
\name{popTableIDs}
\alias{popTableIDs}
\title{List of population table unique IDs.}
\usage{
popTableIDs(
  cuniqueid = "PLT_CN",
  puniqueid = "CN",
  tuniqueid = "PLT_CN",
  vuniqueid = "PLT_CN",
  subpuniqueid = "PLT_CN",
  lulcuniqueid = "PLT_CN",
  condid = "CONDID",
  ...
)
}
\arguments{
\item{cuniqueid}{String. Unique identifier of plot in cond.}

\item{puniqueid}{String. Unique identifier of plot in plt.}

\item{tuniqueid}{String. Unique identifier of plot in tree and seed.}

\item{vuniqueid}{String.}

\item{subpuniqueid}{String.}

\item{lulcuniqueid}{String.}

\item{condid}{String.}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for strata.
}
\description{
Returns a list of user-supplied parameters and parameter values for data 
table unique IDs to be supplied to *pop functions.
}
\details{
If no parameters, an empty list is returned.
}
\examples{

popTableIDs()

}
\author{
Grayson W. White
}
\keyword{list}
