% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{class}
\name{Rcpp_CmbTable-class}
\alias{Rcpp_CmbTable-class}
\title{Class \code{"Rcpp_CmbTable"}}
\description{
C++ program to combine raster files.
}
\section{Extends}{
 Class \code{"\linkS4class{C++Object}"}, directly.

All reference classes extend and inherit methods from
\code{"\linkS4class{envRefClass}"}.
}

\author{
Chris Toney
}
\keyword{classes}
