\name{Rcpp_RunningStats-class}
\Rdversion{1.1}
\docType{class}
\alias{Rcpp_RunningStats-class}

\title{Class \code{"Rcpp_RunningStats"}}
\description{
  C++ program to calculate mean and variance on a data stream.
}
\section{Extends}{
Class \code{"\linkS4class{C++Object}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.

}
\author{
  Chris Toney
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\keyword{classes}
\section{Methods}{
  \describe{
    \item{\code{initialize(...)}:}{ ~~ }
    \item{\code{finalize()}:}{ ~~ }
    \item{\code{update(...)}:}{ ~~ }
    \item{\code{reset()}:}{ ~~ }
    \item{\code{get_sum()}:}{ ~~ }
    \item{\code{get_sampVar()}:}{ ~~ }
    \item{\code{get_sampSD()}:}{ ~~ }
    \item{\code{get_popVar()}:}{ ~~ }
    \item{\code{get_popSD()}:}{ ~~ }
    \item{\code{get_min()}:}{ ~~ }
    \item{\code{get_mean()}:}{ ~~ }
    \item{\code{get_max()}:}{ ~~ }
    \item{\code{get_count()}:}{ ~~ }
  }
}

