\name{ref_actualc}
\docType{data}
\alias{ref_actualc}
\title{ Reference table - Metadata for actualc default variables output from DBgetPlots() }
\description{ Data frame with variable names and descriptions }
\usage{ ref_actualc }

\format{ 
A data frame with 5 rows and 2 columns 
	VARIABLE - Variable in actualc data frame
	DESCRIPTION - Description of variable in actualc data frame
}
\source{ FIA look-up table }
\references{
	O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer, D.; Wilson, A.M.; Waddell, K.L.;
	Christensen, G.; Conkling, B.L. 2012. The Forest Inventory and Analysis Database: Database
	Description and Users Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture. 
	(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user manual_5-1-2_p2_07_2012.pdf)
}
\keyword{ datasets }
