% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_codes}
\alias{ref_codes}
\title{Reference tables - Code definitions.}
\format{
A dataframe with 7 columns, VARIABLE, VALUE, MEANING, COLORHEX,
GROUP, GROUPNM, GROUPHEX.
}
\source{
FIA look-up tables.
}
\description{
Table with variable codes (VALUE) and descriptions (MEANING).
}
\references{
O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer,
D.; Wilson, A.M.; Waddell, K.L.; Christensen, G.; Conkling, B.L. 2012. The
Forest Inventory and Analysis Database: Database Description and Users
Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture.
(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user
manual_5-1-2_p2_07_2012.pdf)
}
\keyword{datasets}
