\name{ref_conversion}
\docType{data}
\alias{ref_conversion}
\title{ Reference table - for generating tables. }
\description{ Table conversion factors from English to metric units. }
\usage{ ref_conversion }
\format{ A dataframe with 4 columns: TYPE, ENGLISH, METRIC, CONVERSION. }
\source{ Conversion table.  }

\keyword{ datasets }
