\name{ref_diacl2in}
\docType{data}
\alias{ref_diacl2in}
\title{ Reference table - diameter 2-inch class codes (DIA). }
\description{ Table with min (MIN), max (MAX), and 2-inch class diameter codes (MEANING). }
\usage{ ref_diacl2in }
\format{ A dataframe with 3 columns, MIN, MAX, and MEANING. }
\source{ Imported from comma-delimited file.  }

\references{
	O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer, D.; Wilson, A.M.; Waddell, K.L.;
	Christensen, G.; Conkling, B.L. 2012. The Forest Inventory and Analysis Database: Database
	Description and Users Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture. 
	(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user manual_5-1-2_p2_07_2012.pdf)
}
\keyword{ datasets }
