\name{ref_domain}
\docType{data}
\alias{ref_domain}
\title{ Reference table - for generating tables. }
\description{ Table with row/column domain (VARNM) and their pretty names for table output (TABLENM). }
\usage{ ref_domain }
\format{ A dataframe with 2 columns, VARNM and TABLENM. }
\source{ FIA look-up table.  }

\keyword{ datasets }
