% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_estvar}
\alias{ref_estvar}
\title{Reference table - for generating tables}
\format{
An object of class \code{data.frame} with 39 rows and 8 columns.
}
\source{
FIA look-up table
}
\usage{
ref_estvar
}
\description{
Reference table - for generating tables
}
\keyword{datasets}
