\name{ref_ICE_agent}
\docType{data}
\alias{ref_ICE_agent}
\title{ Reference table - ICE - land use or land cover change. }
\description{ Table with land use or land cover change codes and descriptions. }
\usage{ ref_ICE_agent }
\format{ A dataframe with 6 columns: chg_ag_2, chg_ag_2_nm, chg_ag_2_GRP, chg_ag_2_GRP_nm,
		change_pnt, change_pnt_nm. }
\source{ Imported from comma-delimited file.  }

\keyword{ datasets }
