\name{ref_ICE_change}
\docType{data}
\alias{ref_ICE_change}
\title{ Reference table - ICE - land use or land cover change. }
\description{ Table with land use or land cover change codes and descriptions. }
\usage{ ref_ICE_change }
\format{ A dataframe with 4 columns: change_1_2, change_1_2_nm, change_1_2_GRP, change_1_2_GRP_nm. }
\source{ Imported from comma-delimited file.  }

\keyword{ datasets }
