% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_ICE_cover}
\alias{ref_ICE_cover}
\title{Reference table - ICE - land use cover.}
\format{
A dataframe with 7 columns: cover, cover_nm, cover_GRP,
cover_GRP_nm, cover_GRP2, cover_GRP2_nm, COLOR.
}
\source{
Imported from comma-delimited file.
}
\description{
Table with land cover codes and descriptions.
}
\keyword{datasets}
