\name{ref_ICE_domain}
\docType{data}
\alias{ref_ICE_domain}
\title{ Reference table - for generating tables. }
\description{ Table with row/column domain (VARNM) and their pretty names for table output (TABLENM). }
\usage{ ref_ICE_domain }
\format{ A dataframe with 3 columns, DOMCODE, DOMNAME, DOMTITLE. }
\source{ FIA ICE look-up table.  }

\keyword{ datasets }
