% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_ICE_use}
\alias{ref_ICE_use}
\title{Reference table - ICE - land use.}
\format{
A dataframe with 4 columns: use, use_nm, use_FOR, use_FOR_nm.
}
\source{
Imported from comma-delimited file.
}
\description{
Table with land use codes and descriptions.
}
\keyword{datasets}
