\name{ref_icepnt}
\docType{data}
\alias{ref_icepnt}
\title{ Reference table - Metadata for icepnt external dataset. }
\description{ Data frame with variable names and descriptions. }
\usage{ ref_icepnt }

\format{ A data frame with 8 rows and 2 columns
	VARIABLE - Variable in plt data frame
	DESCRIPTION - Description of variable in data frame
}
\source{ FIA look-up table }
\keyword{ datasets }
