% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_popType}
\alias{ref_popType}
\title{Reference table - popType codes.}
\format{
A dataframe with 2 columns, VARNM and TITLE.
}
\source{
Comma-delimited file.
}
\description{
Table with population type (popType) and associated evaluation code (EVAL_TYP_CD).
}
\keyword{datasets}
