\name{ref_species}
\docType{data}
\alias{ref_species}
\title{ Reference tables - Code definitions. }
\description{ Table with species information. }
\usage{ ref_species}
\format{ A dataframe with 8 columns: SPCD, COMMON_NAME, GENUS, SPECIES, 
	EXISTS_IN_NRCS, EXISTS_IN_NERS, EXISTS_IN_PNWRS, EXISTS_IN_RMRS, EXISTS_IN_SRS. }
\source{ FIA look-up tables.  }

\references{
	O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer, D.; Wilson, A.M.; Waddell, K.L.;
	Christensen, G.; Conkling, B.L. 2012. The Forest Inventory and Analysis Database: Database
	Description and Users Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture. 
	(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user manual_5-1-2_p2_07_2012.pdf)
}
\keyword{ datasets }
