% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_stdagecl}
\alias{ref_stdagecl}
\title{Reference table - 20-year stand age classes(STDAGE).}
\format{
A dataframe with 2 columns, MIN, and STDAGENM.
}
\source{
Imported from comma-delimited file.
}
\description{
Table with min (MIN), and stand age class name (STDAGECLNM).
}
\references{
O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer,
D.; Wilson, A.M.; Waddell, K.L.; Christensen, G.; Conkling, B.L. 2012. The
Forest Inventory and Analysis Database: Database Description and Users
Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture.
(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user
manual_5-1-2_p2_07_2012.pdf)
}
\keyword{datasets}
