\name{ref_titles}
\docType{data}
\alias{ref_titles}
\title{ Reference table - Variable titles. }
\description{ Table with variable name (VARNM) and associated title (TITLE). }
\usage{ ref_titles }
\format{ A dataframe with 2 columns, VARNM and TITLE. }
\source{ Comma-delimited file.  }

\keyword{ datasets }
