% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_tree}
\alias{ref_tree}
\title{Reference table - Metadata for tree default variables output from
DBgetPlots()}
\format{
An object of class \code{data.frame} with 117 rows and 3 columns.
}
\source{
FIA look-up table
}
\usage{
ref_tree
}
\description{
Reference table - Metadata for tree default variables output from
DBgetPlots()
}
\keyword{datasets}
