% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_units}
\alias{ref_units}
\title{Reference table - for variable units.}
\format{
An object of class \code{data.frame} with 47 rows and 5 columns.
}
\source{
Units table.
}
\usage{
ref_units
}
\description{
Reference table - for variable units.
}
\keyword{datasets}
