% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spExportSpatial.R
\name{spExportSpatial}
\alias{spExportSpatial}
\title{Spatial - Exports an sf object.}
\usage{
spExportSpatial(sfobj, savedata_opts = savedata_options())
}
\arguments{
\item{sfobj}{sf class R object. Spatial object to export.}

\item{savedata_opts}{List. See help(savedata_options()) for a list
of options.}
}
\value{
An sf spatial object is written to the out_dsn.
}
\description{
Exports an sf object to a specified output.
}
\details{
Wrapper for sf::st_write function.
}
\note{
If out_fmt='shp':\cr The ESRI shapefile driver truncates variable
names to 10 characters or less.  Variable names are changed before export
using an internal function (trunc10shp). Name changes are output to the
outfolder, 'outshpnm'_newnames.csv.

If sf object has more than 1 record, it cannot be exported to a shapefile.
}
\author{
Tracey S. Frescino
}
\keyword{data}
