% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_options.R
\name{unit_options}
\alias{unit_options}
\title{Unit options.}
\usage{
unit_options(
  dunitvar2 = NULL,
  areavar = "ACRES",
  areaunits = "acres",
  minplotnum.unit = 10,
  unit.action = "keep",
  npixelvar = "npixels",
  ...
)
}
\arguments{
\item{dunitvar2}{}

\item{areavar}{}

\item{minplotnum.unit}{}

\item{unit.action}{}

\item{npixelvar}{}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for strata.
}
\description{
Returns a list of user-supplied parameters and parameter values for unit.
}
\details{
If no parameters, an empty list is returned.
}
\examples{

unit_options()

}
\author{
Grayson W. White
}
\keyword{options}
