% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{WYplt}
\alias{WYplt}
\title{FIA data. Plot-level data from FIA public database.}
\format{
A data frame with 20 columns and 3047 rows.
}
\source{
FIA national database (FIADB_1.7.0.00), downloaded September 18, 2016.
}
\description{
FIA plot-level data for the state of Wyoming, FIA Evaluation 561301, 
including inventory years 2011-2013.
}
\details{
\tabular{lll}{ \tab \bold{VARIABLE} \tab \bold{DESCRIPTION} \cr \tab CN \tab
Unique FIADB identifier \cr \tab PREV_PLT_CN \tab Previous unique FIADB
identifier \cr \tab INVYR \tab Inventory year \cr \tab STATECD \tab State
code (FIPS) \cr \tab CYCLE \tab Inventory cycle number \cr \tab SUBCYCLE
\tab Inventory subcycle number (Do not use subcycle 99 for estimation) \cr
\tab UNITCD \tab Survey unit code \cr \tab COUNTYCD \tab County code \cr
\tab PLOT \tab Phase 2 plot number (Public) \cr \tab LON_PUBLIC \tab
Longitude - fuzzed/swapped (Decimal degrees; NAD83) \cr \tab LAT_PUBLIC \tab
Latitude - fuzzed/swapped (Decimal degrees; NAD83) \cr \tab
PLOT_NONSAMPLE_REASN_CD \tab Plot nonsampled reason \cr \tab SAMP_METHOD_CD
\tab Sample method code \cr \tab SUBP_EXAMINE_CD \tab Subplots examined code
\cr \tab MANUAL \tab Manual version number \cr \tab INTENSITY \tab Intensity
\cr \tab MEASYEAR \tab Measurement year \cr \tab MEASMON \tab Measurement
month \cr \tab MEASDAY \tab Measurement day \cr \tab REMPER \tab
Remeasurement period \cr \tab DESIGNCD \tab Plot design \cr \tab P2PANEL
\tab Phase 2 panel number \cr \tab SUBPANEL \tab Subpanel number \cr \tab
ELEV \tab Elevation (ft) \cr \tab KINDCD \tab Sample kind \cr \tab
MORT_TYP_CD \tab Type of annual mortality volume (1:Current annual;
2:Periodic annual) \cr \tab GROW_TYP_CD \tab Type of annual volume growth
(1:Current annual; 2:Periodic annual) \cr \tab NF_PLOT_NONSAMPLE_REASN_CD
\tab Nonforest sampling status \cr \tab P2VEG_SAMPLING_STATUS_CD \tab P2
vegetation sampling status \cr \tab PLOT_STATUS_CD \tab Plot sampling status
\cr \tab NF_PLOT_STATUS_CD \tab Nonforest plot sampling status \cr \tab
NBRCND \tab DERIVED: Number of conditions for plot \cr \tab NBRCNDFOR \tab
DERIVED: Number of different forest type conditions for plot \cr \tab
FORNONSAMP \tab DERIVED: Plot status - sampled and nonsampled (Combination
of PLOT_NONSAMPLE_REASN_CD and PLOT_STATUS_CD) \cr \tab CCLIVEPLT \tab
DERIVED: Percent cover of live trees for plot (LIVE_CANOPY_CVR_PCT *
CONDPROP_UNADJ) \cr \tab UNIQUEID \tab DERIVED: Unique identifier for plot
location ('Z'+STATECD(2)+UNITCD(2)+COUNTYCD(3)+PLOT(5)) \cr }
}
\references{
Burrill, E.A.; Wilson, A.M.; Turner, J.A.; Pugh, S.A.; Menlove, J.; 
Christiansen, G.; B.L. Conkling, B.L.: David, W. 2018. The Forest Inventory and 
Analysis Database: Database description and user guide version 8.0 for Phase 2.
}
\keyword{datasets}
