% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{WYpltassgn}
\alias{WYpltassgn}
\title{FIA data. Plot-level data from FIA public database.}
\format{
A dataframe with 24 columns and 3047 rows.
}
\source{
FIA Database.
}
\description{
FIA plot-level data for the state of Wyoming, inventory years 2011-2012 and
was downloaded Sept 8, 2017 from FIADB_1.7.0.00. It includes strata
assignments for FIA plot center location (1:GREEN; 2:BROWN).
}
\references{
O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer,
D.; Wilson, A.M.; Waddell, K.L.; Christensen, G.; Conkling, B.L. 2012. The
Forest Inventory and Analysis Database: Database Description and Users
Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture.
(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user
manual_5-1-2_p2_07_2012.pdf)
}
\keyword{datasets}
