% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{WYseed}
\alias{WYseed}
\title{FIA data. Seedling data from FIA public database.}
\format{
A dataframe with 10 columns and 1039 rows.
}
\source{
FIA Database.
}
\description{
FIA seedling data for the state of Wyoming, inventory years 2011-2012 and
was downloaded May 17, 2017 from FIADB_1.7.0.00.
}
\references{
O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer,
D.; Wilson, A.M.; Waddell, K.L.; Christensen, G.; Conkling, B.L. 2012. The
Forest Inventory and Analysis Database: Database Description and Users
Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture.
(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user
manual_5-1-2_p2_07_2012.pdf)
}
\keyword{datasets}
