% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{WYstratalut}
\alias{WYstratalut}
\title{FIA data. Strata data from FIA public database.}
\format{
A dataframe with 6 columns and 39 rows.
}
\source{
FIA Database.
}
\description{
FIA condition-level data for the state of Wyoming, inventory years 2011-2012
and was downloaded Sept 8, 2017 from FIADB_1.7.0.00. It includes P1POINTCNT
and number of FIA plots by strata (1:GREEN; 2:BROWN) and by estimation unit
(ESTN_UNIT).
}
\references{
O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer,
D.; Wilson, A.M.; Waddell, K.L.; Christensen, G.; Conkling, B.L. 2012. The
Forest Inventory and Analysis Database: Database Description and Users
Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture.
(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user
manual_5-1-2_p2_07_2012.pdf)
}
\keyword{datasets}
