% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{WYsubplot}
\alias{WYsubplot}
\title{FIA data. Subplot data from FIA public database.}
\format{
A dataframe with 9 columns and 20596 rows.
}
\source{
FIA national database (FIADB_1.7.0.00), downloaded September 18, 2016.
}
\description{
FIA subplot-level data for the state of Wyoming, FIA Evaluation 561301, 
including inventory years 2011-2013.
}
\references{
Burrill, E.A.; Wilson, A.M.; Turner, J.A.; Pugh, S.A.; Menlove, J.; 
Christiansen, G.; B.L. Conkling, B.L.: David, W. 2018. The Forest Inventory and 
Analysis Database: Database description and user guide version 8.0 for Phase 2.
}
\keyword{datasets}
