% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{WYunitzonal}
\alias{WYunitzonal}
\title{Zonal data. Zonal means for auxiliary data in counties in Wyoming.}
\format{
A dataframe with 9 columns and 23 rows.
}
\description{
Zonal means and pixel counts for certain auxiliary data in counties in
Wyoming. Includes county code variable to distinguish counties, and state
code variable to distinguish states.
}
\keyword{datasets}
