% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datPBplotchg.R
\name{datPBplotchg}
\alias{datPBplotchg}
\title{Data - Generates barplot of photo-based change estimates.}
\usage{
datPBplotchg(gainloss, CI = 95, figTitle = "")
}
\arguments{
\item{gainloss}{Data frame or comma-delimited file (*.csv) - table with gain
loss estimates.}

\item{CI}{Number. Confidence Interval to include on plot.}

\item{figTitle}{String. Title of figure.}
}
\value{
Outputs barplot to display window.
}
\description{
Generate a bar plot of net change for photo-based estimates of land use /
land cover change.
}
\note{
If savedata = TRUE, writes a jpg and pdf of barplot to outfolder.
}
\author{
Tracey S. Frescino
}
\keyword{data}
