% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datSumTreeDomRatio.R
\name{datSumTreeDomRatio}
\alias{datSumTreeDomRatio}
\title{Data - Generates ratio of tree domain variables.}
\usage{
datSumTreeDomRatio(
  ndat,
  ddat,
  uniqueid = "PLT_CN",
  nprefix = NULL,
  dprefix = NULL,
  rprefix = "r",
  datround = NULL,
  savedata = FALSE,
  savedata_opts = NULL,
  gui = FALSE
)
}
\arguments{
\item{ndat}{Dataframe, comma-delimited file (*.csv), or shapefile (*.shp).
Table from FIESTA::datSumTreeDomRatio() with numerator tree domain
variables.}

\item{ddat}{Dataframe, comma-delimited file (*.csv), or shapefile (*.shp).
Table from FIESTA::datSumTreeDomRatio() with numerator tree domain
variables.}

\item{uniqueid}{String. The unique identifier of both tables (default =
"PLT_CN").}

\item{nprefix}{String. The prefix variable identifier of numerator domain
variables in ndat.}

\item{dprefix}{String. The prefix variable identifier of denominator domain
variables in ddat.}

\item{rprefix}{String. The prefix variable identifier of new ratio variables
(default="r").}

\item{datround}{Integer. Number of digits to round ratio values to.}

\item{savedata}{Logical. If TRUE, saves data to outfolder.}

\item{savedata_opts}{List. See help(savedata_options()) for a list
of options. Only used when savedata = TRUE. If out_layer = NULL,
default = 'tsumrat'.}

\item{gui}{Logical. If gui, user is prompted for parameters.}
}
\value{
\item{rdat}{ Data frame. Table with ratio values (ndat / ddat). }
\item{rvars}{ String vector. Variable names in rdat. }

If savedata=TRUE, the data table will be saved to the outfolder: \cr
}
\description{
Generates ratio of tree domain summaries from FIESTA::datSumTreedom().
}
\examples{
# Ratio of live and dead basal area
# Live basal area
live <- datSumTreeDom(tree = FIESTA::WYtree, 
                      tuniqueid = "PLT_CN", 
                      tsumvar = "BA", 
                      tfilter = "STATUSCD==1")
                      
# Dead basal area                       
dead <- datSumTreeDom(tree = FIESTA::WYtree, 
                      tuniqueid = "PLT_CN", 
                      tsumvar = "BA", 
                      tfilter = "STATUSCD == 2 & STANDING_DEAD_CD == 1")
                      
FIESTA:::datSumTreeDomRatio(ndat = dead$tdomdat,
                            ddat = live$tdomdat)             
}
\author{
Tracey S. Frescino
}
\keyword{data}
