% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBgetSQLite.R
\name{DBgetSQLite}
\alias{DBgetSQLite}
\title{Database - Queries a SQLite database table.}
\usage{
DBgetSQLite(states = NULL, outfolder = NULL)
}
\arguments{
\item{states}{String. Vector of one or more state names.}

\item{outfolder}{String. The output folder path. If NULL, outfolder is the
working directory.}
}
\value{
Returns nothing.
}
\description{
Extracts and queries data from a SQLite (*.sqlite) database (Note: must use
SQL syntax).
}
\examples{
\dontrun{
# Extract data from Washington state
DBgetSQLite(states = "WA")

# Extract data from Utah and California, save to an outfolder
DBgetSQLite(states = c("UT", "CA"),
            outfolder = tempdir()) 
}
}
\author{
Tracey S. Frescino
}
\keyword{data}
