\name{FIESTA-internal}
\alias{addCI}
\alias{addcommas}
\alias{DBvars.default}
\alias{change10char}
\alias{check.logic}
\alias{check.estdata}
\alias{check.popdata}
\alias{check.estdataPB}
\alias{check.popdataPB}
\alias{check.nostrata}
\alias{check.auxiliary}
\alias{check.rowcol}
\alias{check.rowcolPB}
\alias{check.titles}
\alias{check.tree}
\alias{check.PROP}
\alias{check.unitarea}
\alias{check.matchclass}
\alias{check.matchval}
\alias{check.numeric}
\alias{check.pltcnt}
\alias{check.extents}
\alias{fileexistsnm}
\alias{est.outtabs}
\alias{getarea}
\alias{getadjfactorGB}
\alias{getadjfactorPLOT}
\alias{getcrossord}
\alias{getfunnm}
\alias{getoutfn}
\alias{getlt10char}
\alias{getRefobject}
\alias{getrhat}
\alias{getspconddat}
\alias{getstratacnt}
\alias{groupEstunit}
\alias{groupStrata}
\alias{groupUnits}
\alias{makebrick}
\alias{pastevars}
\alias{pcheck.DBtabs}
\alias{pcheck.dsn}
\alias{pcheck.landarea}
\alias{pcheck.logical}
\alias{pcheck.outfolder}
\alias{pcheck.rowcol}
\alias{pcheck.states}
\alias{pcheck.strata}
\alias{pcheck.table}
\alias{pcheck.unitacres}
\alias{pcheck.varchar}
\alias{pcheck.xlsx}
\alias{projfix}
\alias{save1tab}
\alias{save2tabs}
\alias{strat.collapse}
\alias{write2csv}
\alias{write2gdb}
\alias{write2sqlite}
\alias{wtcalc}
\alias{wtcalc.unit}
\alias{bp_wrap.it}
\alias{bp_width.jpg}

\alias{stopQ}
\alias{removecols}
\alias{DT_NAto0}
\alias{changeNULL}
\alias{getdups}
\alias{getlistparam}
\alias{getnm}
\alias{checknm}
\alias{check.namedlist}
\alias{capfirst}
\alias{DBgetfn}
\alias{getpfromqry}
\alias{getEvalid}
\alias{getEvalid.ppsa}
\alias{getPlotCur}
\alias{nbrdecimals}
\alias{nbrdigits}
\alias{getfilter}

\alias{PBest.pbar}
\alias{PBest.pbarRatio}
\alias{PBgetest}
\alias{getpltdom.prop}
\alias{getgainloss}
\alias{transpose2row}
\alias{transpose2col}

\alias{add0unit}
\alias{addrowgrp}
\alias{allin1f}
\alias{crosstabx}
\alias{crossxtab}
\alias{getdomain}
\alias{getestvar}

\alias{wrapSE}
\alias{tabgrp}
\alias{setCells}
\alias{write2xlsx}
\alias{wraptitle}

\alias{getext}
\alias{getrastlst.rgdal}
\alias{polyfix.sf}
\alias{pcheck.spatial}
\alias{build.prj4str}
\alias{trunc10shp}
\alias{getEPSG}
\alias{check.extents} 
\alias{getprjatt}
\alias{areacalc.poly}
\alias{areacalc.pixel}
\alias{aspect_transform}
\alias{spPlotRastcl}
\alias{checksf.longlat}
\alias{checkrast.longlat}
\alias{crsCompare}
\alias{sf_dissolve}
\alias{closest_poly}
\alias{getIntersect}
\alias{clip.othertables}
\alias{spGetStates}

\alias{getOffset}
\alias{getGDALDataTypeName}
\alias{getGDALformat}
\alias{basename.NoExt}
\alias{Mode}
\alias{northness}
\alias{eastness}
\alias{getPixelValue}
\alias{extractPtsFromRaster}
\alias{extractPtsFromRasterList}
\alias{rasterInfo}
\alias{reprojectRaster}
\alias{rasterFromRaster}
\alias{rasterFromVectorExtent}
\alias{RasterizePolygon}
\alias{rasterizePolygons}
\alias{polygonizeRaster}
\alias{pixelCount}
\alias{clipRaster}
\alias{rasterCombine}
\alias{recodeRaster}
\alias{zonalMean}
\alias{zonalFreq}
\alias{zonalMajority}
\alias{zonalMinority}
\alias{zonalVariety}
\alias{ptCsvToVRT}
\alias{CmbTable} 
\alias{RunningStats}
\alias{getGDALformat}
\alias{ptCsvToVRT}
\alias{rasterToVRT}
\alias{zonalStats}
\alias{DEFAULT_NODATA}
\alias{GDT_NAMES}
\alias{getDefaultNodata}
\alias{rasterCalc}
\alias{TPI}
\alias{TRI}
\alias{focalRaster}
\alias{roughness}


\alias{ptdata2sf}
\alias{sf2ptdata}
\alias{layerIntersection}
\alias{layerUnion}
\alias{tabulateIntersections}
\alias{selectByIntersects}
\alias{layerFromRasterExtent}
\alias{vectorTranslate}


\alias{GBest.pbar}
\alias{GBest.pop}

\alias{MAest.pop}
\alias{MAest.unit}
\alias{MAest.dom}
\alias{MAest}
\alias{MAest.ps}
\alias{MAest.greg}
\alias{MAest.ht}

\alias{SAest.pop}
\alias{SAest.large}
\alias{SAest.dom}
\alias{SAest}
\alias{FIESTA_SAmod_demo_plots}
\alias{helper.select}



\title{ Internal FIESTA Functions }
\description{
  Internal Subfunctions
}

\details{
These are not to be called by the user.  
}

\author{ Tracey S. Frescino }

\keyword{internal}

