\name{FIESTA-internal}
\alias{getRefcodes}
\alias{strat.pivot}
\alias{preds.standardize}
\alias{preds.select}
\alias{spGetStates}
\alias{rasterCombine}
\alias{CmbTable} 
\alias{rasterInfo}
\alias{rasterFromRaster}
\alias{rasterFromVectorExtent}
\alias{rasterizePolygons}
\alias{rasterCalc}
\alias{focalRaster}
\alias{recodeRaster}
\alias{polygonizeRaster}
\alias{getPixelValue}


\title{ Internal FIESTA Functions }
\description{
  Internal Subfunctions
}

\details{
These are not to be called by the user.  
}

\author{ Tracey S. Frescino }

\keyword{internal}

