% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_description.R
\name{internal_desciption}
\alias{internal_desciption}
\title{Internal functions used for development of FIESTA}
\description{
These functions are used internally by developers of the FIESTA and
FIESTAnalysis R packages. They are not meant to be used on their own, or by
users. While some functions may be useful on their own, they will not be
regularly updated, documented, or changed for a particular use.
}
\details{
These functions are not to be called by a user, and are for development
purposes only.
}
\author{
Tracey S. Frescino, Chris Toney, Grayson W. White
}
\keyword{internal}
