% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popTableIDs.R
\name{popTableIDs}
\alias{popTableIDs}
\title{List of population table unique IDs.}
\usage{
popTableIDs(
  cond = "PLT_CN",
  plt = "CN",
  tree = "PLT_CN",
  seed = "PLT_CN",
  vsubpspp = "PLT_CN",
  vsubpstr = "PLT_CN",
  invsubp = "PLT_CN",
  subplot = "PLT_CN",
  subp_cond = "PLT_CN",
  cond_dwm_calc = "PLT_CN",
  sccm = "PLT_CN",
  grm = "PLT_CN",
  pltu = "PLT_CN",
  condu = "PLT_CN",
  begin = "PLT_CN",
  midpt = "PLT_CN",
  ...
)
}
\arguments{
\item{cond}{String. Unique identifier of plot in cond.}

\item{plt}{String. Unique identifier of plot in plt.}

\item{tree}{String. Unique identifier of plot in tree and seed.}

\item{seed}{String.}

\item{vsubpspp}{String.}

\item{vsubpstr}{String.}

\item{invsubp}{String.}

\item{subplot}{String.}

\item{subp_cond}{String.}

\item{cond_dwm_calc}{String.}

\item{sccm}{String.}

\item{grm}{String.}

\item{pltu}{String.}

\item{condu}{String.}

\item{begin}{String.}

\item{midpt}{String.}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for strata.
}
\description{
Returns a list of user-supplied parameters and parameter values for data 
table unique IDs to be supplied to *pop functions.
}
\details{
If no parameters, an empty list is returned.
}
\examples{
popTableIDs(cond = "my_unique_id", tree = "my_tree_id")
}
\author{
Grayson W. White
}
\keyword{list}
