\name{Rcpp_CmbTable-class}
\Rdversion{1.1}
\docType{class}
\alias{Rcpp_CmbTable-class}

\title{Class \code{"Rcpp_CmbTable"}}
\description{
  C++ program to combine raster files.
}
\section{Extends}{
Class \code{"\linkS4class{C++Object}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.

}
\author{
  Chris Toney
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\keyword{classes}
\section{Methods}{
  \describe{
    \item{\code{initialize(...)}:}{ ~~ }
    \item{\code{finalize()}:}{ ~~ }
    \item{\code{updateFromMatrix(...)}:}{ ~~ }
    \item{\code{update(...)}:}{ ~~ }
    \item{\code{asDataFrame()}:}{ ~~ }
  }
}
