% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{class}
\name{Rcpp_RunningStats-class}
\alias{Rcpp_RunningStats-class}
\title{Class \code{"Rcpp_RunningStats"}}
\description{
C++ program to calculate mean and variance on a data stream.
}
\section{Extends}{
 Class \code{"\linkS4class{C++Object}"}, directly.

All reference classes extend and inherit methods from
\code{"\linkS4class{envRefClass}"}.
}

\author{
Chris Toney
}
\keyword{classes}
