% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{savedata_options}
\alias{unit_options}
\alias{table_options}
\alias{title_options}
\alias{strata_options}
\alias{multest_options}
\alias{spMakeSpatial_options}
\alias{datExportData}
\alias{DBcreateSQLite}
\alias{DBtestPostgreSQL}
\alias{DBtestSQLite}
\alias{eval_options}
\alias{xy_options}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{FIESTAutils}{\code{\link[FIESTAutils]{datExportData}}, \code{\link[FIESTAutils:internal_desc]{DBcreateSQLite}}, \code{\link[FIESTAutils]{DBtestPostgreSQL}}, \code{\link[FIESTAutils]{DBtestSQLite}}, \code{\link[FIESTAutils]{eval_options}}, \code{\link[FIESTAutils]{multest_options}}, \code{\link[FIESTAutils]{savedata_options}}, \code{\link[FIESTAutils]{spMakeSpatial_options}}, \code{\link[FIESTAutils]{strata_options}}, \code{\link[FIESTAutils]{table_options}}, \code{\link[FIESTAutils]{title_options}}, \code{\link[FIESTAutils]{unit_options}}, \code{\link[FIESTAutils]{xy_options}}}
}}

