% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_codes}
\alias{ref_codes}
\title{Reference tables - Code definitions.}
\format{
An object of class \code{data.frame} with 735 rows and 7 columns.
}
\source{
FIA look-up tables.
}
\usage{
ref_codes
}
\description{
Reference tables - Code definitions.
}
\references{
O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer,
D.; Wilson, A.M.; Waddell, K.L.; Christensen, G.; Conkling, B.L. 2012. The
Forest Inventory and Analysis Database: Database Description and Users
Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture.
(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user
manual_5-1-2_p2_07_2012.pdf)
}
\keyword{datasets}
