% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_conversion}
\alias{ref_conversion}
\title{Reference table - for generating tables.}
\format{
An object of class \code{data.frame} with 7 rows and 4 columns.
}
\source{
Conversion table.
}
\usage{
ref_conversion
}
\description{
Reference table - for generating tables.
}
\keyword{datasets}
