% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_conversion}
\alias{ref_conversion}
\title{Reference table - for generating tables.}
\format{
A dataframe with 4 columns: TYPE, ENGLISH, METRIC, CONVERSION.
}
\source{
Conversion table.
}
\description{
Table conversion factors from English to metric units.
}
\keyword{datasets}
