% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_domain}
\alias{ref_domain}
\title{Reference table - for generating tables.}
\format{
A dataframe with 2 columns, VARNM and TABLENM.
}
\source{
FIA look-up table.
}
\description{
Table with row/column domain (VARNM) and their pretty names for table output
(TABLENM).
}
\keyword{datasets}
