% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_domain}
\alias{ref_domain}
\title{Reference table - for generating tables.}
\format{
An object of class \code{data.frame} with 33 rows and 3 columns.
}
\source{
FIA look-up table.
}
\usage{
ref_domain
}
\description{
Reference table - for generating tables.
}
\keyword{datasets}
