\name{ref_estvar}
\docType{data}
\alias{ref_estvar}
\title{ Reference table - for generating tables }
\description{ Data frame with variable names and descriptions }
\usage{ ref_estvar}

\format{ A data frame with 26 rows and 4 columns
	CATEGORY - Category of estimation variable
	ESVARNM - Estimation variable in database
 	ESTTITLE - A title for estimation selections
 	ESTFILTER - Filter statement for each estimation selection
}
\source{ FIA look-up table  }

\keyword{ datasets }
