% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_estvar}
\alias{ref_estvar}
\title{Reference table - for generating tables}
\format{
A data frame with 26 rows and 4 columns CATEGORY - Category of
estimation variable ESVARNM - Estimation variable in database ESTTITLE - A
title for estimation selections ESTFILTER - Filter statement for each
estimation selection
}
\source{
FIA look-up table
}
\description{
Data frame with variable names and descriptions
}
\keyword{datasets}
