% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_estvar}
\alias{ref_estvar}
\title{Reference table - for generating estimates}
\format{
An object of class \code{data.frame} with 32 rows and 9 columns.
}
\usage{
ref_estvar
}
\description{
Reference table - for generating estimates
}
\keyword{datasets}
