% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_ICE_agent}
\alias{ref_ICE_agent}
\title{Reference table - ICE - land use or land cover change.}
\format{
A dataframe with 6 columns: chg_ag_2, chg_ag_2_nm, chg_ag_2_GRP,
chg_ag_2_GRP_nm, change_pnt, change_pnt_nm.
}
\source{
Imported from comma-delimited file.
}
\description{
Table with land use or land cover change codes and descriptions.
}
\keyword{datasets}
