% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_ICE_change}
\alias{ref_ICE_change}
\title{Reference table - ICE - land use or land cover change.}
\format{
A dataframe with 4 columns: change_1_2, change_1_2_nm,
change_1_2_GRP, change_1_2_GRP_nm.
}
\source{
Imported from comma-delimited file.
}
\description{
Table with land use or land cover change codes and descriptions.
}
\keyword{datasets}
