\name{ref_ICE_cover}
\docType{data}
\alias{ref_ICE_cover}
\title{ Reference table - ICE - land use cover. }
\description{ Table with land cover codes and descriptions. }
\usage{ ref_ICE_cover }
\format{ A dataframe with 7 columns: cover, cover_nm, cover_GRP, cover_GRP_nm, 
		cover_GRP2, cover_GRP2_nm, COLOR. }
\source{ Imported from comma-delimited file.  }

\keyword{ datasets }
