% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_ICE_domain}
\alias{ref_ICE_domain}
\title{Reference table - for generating tables.}
\format{
A dataframe with 3 columns, DOMCODE, DOMNAME, DOMTITLE.
}
\source{
FIA ICE look-up table.
}
\description{
Table with row/column domain (VARNM) and their pretty names for table output
(TABLENM).
}
\keyword{datasets}
