\name{ref_ICE_use}
\docType{data}
\alias{ref_ICE_use}
\title{ Reference table - ICE - land use. }
\description{ Table with land use codes and descriptions. }
\usage{ ref_ICE_use }
\format{ A dataframe with 4 columns: use, use_nm, use_FOR, use_FOR_nm. }
\source{ Imported from comma-delimited file.  }

\keyword{ datasets }
