% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_icepnt}
\alias{ref_icepnt}
\title{Reference table - Metadata for icepnt external dataset.}
\format{
A data frame with 8 rows and 2 columns VARIABLE - Variable in plt
data frame DESCRIPTION - Description of variable in data frame
}
\source{
FIA look-up table
}
\description{
Data frame with variable names and descriptions.
}
\keyword{datasets}
