% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_popType}
\alias{ref_popType}
\title{Reference table - popType codes.}
\format{
An object of class \code{data.frame} with 14 rows and 2 columns.
}
\source{
Comma-delimited file.
}
\usage{
ref_popType
}
\description{
Reference table - popType codes.
}
\keyword{datasets}
