% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_shp}
\alias{ref_shp}
\title{Reference table - Metadata for shp_* default variables output from
DBgetPlots()}
\format{
An object of class \code{data.frame} with 63 rows and 4 columns.
}
\source{
FIA look-up table
}
\usage{
ref_shp
}
\description{
Reference table - Metadata for shp_* default variables output from
DBgetPlots()
}
\keyword{datasets}
