% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_species}
\alias{ref_species}
\title{Reference table - Code definitions.}
\format{
An object of class \code{data.frame} with 2677 rows and 14 columns.
}
\source{
Imported from comma-delimited file.
}
\usage{
ref_species
}
\description{
Reference table - Code definitions.
}
\keyword{datasets}
