\name{ref_statecd}
\docType{data}
\alias{ref_statecd}
\title{ Reference table - state codes (STATECD). }
\description{ Table with state codes (VALUE), name (MEANING), abbreviation (ABBR), and UNIT. }
\usage{ ref_statecd }
\format{ A dataframe with 4 columns, VALUE, MEANING, ABBR, UNIT. }
\source{ Imported from comma-delimited file.  }

\references{
	O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer, D.; Wilson, A.M.; Waddell, K.L.;
	Christensen, G.; Conkling, B.L. 2012. The Forest Inventory and Analysis Database: Database
	Description and Users Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture. 
	(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user manual_5-1-2_p2_07_2012.pdf)
}
\keyword{ datasets }
