% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_titles}
\alias{ref_titles}
\title{Reference table - Variable titles.}
\format{
A dataframe with 2 columns, VARNM and TITLE.
}
\source{
Comma-delimited file.
}
\description{
Table with variable name (VARNM) and associated title (TITLE).
}
\keyword{datasets}
