% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{ref_titles}
\alias{ref_titles}
\title{Reference table - Variable titles.}
\format{
An object of class \code{data.frame} with 69 rows and 2 columns.
}
\source{
Comma-delimited file.
}
\usage{
ref_titles
}
\description{
Reference table - Variable titles.
}
\keyword{datasets}
