% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTA-package.R
\docType{data}
\name{ref_tree}
\alias{ref_tree}
\title{Reference table - Metadata for tree default variables output from
DBgetPlots()}
\format{
A data frame with 72 rows and 3 columns VARIABLE - Variable in tree
data frame DESCRIPTION - Description of variable in tree data frame TABLE -
Table in database where variable originates
}
\source{
FIA look-up table
}
\description{
Data frame with variable names and descriptions
}
\references{
O'Connell, B.M.; LaPoint, E.B.; Turner, J.A.; Ridley, T.; Boyer,
D.; Wilson, A.M.; Waddell, K.L.; Christensen, G.; Conkling, B.L. 2012. The
Forest Inventory and Analysis Database: Database Description and Users
Manual Version 5.1.2 for Phase 2. U.S. Department of Agriculture.
(http://fia.fs.fed.us/library/database-documentation/current/ver5-2012/FIADB_user
manual_5-1-2_p2_07_2012.pdf)
}
\keyword{datasets}
