% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spExportSpatial.R
\name{spExportSpatial}
\alias{spExportSpatial}
\title{Spatial - Exports an sf object.}
\usage{
spExportSpatial(
  sfobj,
  out_layer = NULL,
  out_fmt = "shp",
  outfolder = NULL,
  out_dsn = NULL,
  outfn.pre = NULL,
  outfn.date = FALSE,
  overwrite_dsn = FALSE,
  overwrite_layer = TRUE,
  add_layer = TRUE,
  append_layer = FALSE
)
}
\arguments{
\item{sfobj}{sf class R object. Spatial object to export.}

\item{out_layer}{String. Name of layer in out_dsn. If NULL, basename of
out_dsn is used.}

\item{out_fmt}{String. File format for output ('sqlite','gpkg','shp').}

\item{outfolder}{String. Optional. Name of output folder. If NULL, export to
working directory.}

\item{out_dsn}{String. Data source name for output. If extension is not
included, out_fmt is used. Use full path if outfolder=NULL.}

\item{outfn.pre}{String. Prefix for out_dsn.}

\item{outfn.date}{Logical. If TRUE, add current date to out_dsn.}

\item{overwrite_dsn}{Logical. If TRUE, overwrites the out_dsn, if exists.}

\item{overwrite_layer}{Logical. If TRUE, overwrites the out_layer, if
exists.}

\item{add_layer}{Logical. If TRUE, adds to existing out_dsn (if out_fmt !=
c('csv','shp')).}

\item{append_layer}{Logical. If TRUE, appends to existing out_dsn. The
out_dsn a database or shapefile.}
}
\value{
An sf spatial object is written to the out_dsn.
}
\description{
Exports an sf object to a specified output.
}
\details{
Wrapper for sf::st_write function.
}
\note{
If out_fmt='shp':\cr The ESRI shapefile driver truncates variable
names to 10 characters or less.  Variable names are changed before export
using an internal function (trunc10shp). Name changes are output to the
outfolder, 'outshpnm'_newnames.csv.

If sf object has more than 1 record, it cannot be exported to a shapefile.
}
\author{
Tracey S. Frescino
}
\keyword{data}
