% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_options.R
\name{strata_options}
\alias{strata_options}
\title{Strata options.}
\usage{
strata_options(
  stratalut = NULL,
  strvar = "STRATUMCD",
  getwt = TRUE,
  getwtvar = "P1POINTCNT",
  strwtvar = "strwt",
  stratcombine = TRUE,
  minplotnum.strat = 2,
  ...
)
}
\arguments{
\item{stratalut}{DF/DT. If strata=TRUE, look-up table with pixel counts or
area by strata or proportion or area ('strwt') by strata (and estimation
unit).  If 'strwt' is not included, set getwt=TRUE and getwtvar as the name
of variable to calculate weights from (e.g., pixel counts).}

\item{strvar}{String. If strata=TRUE, name of the strata variable in
stratalut and cond or pltassgn data frame with stratum assignment for each
plot (Default = 'STRATUMCD').}

\item{getwt}{Logical. If TRUE, calculates strata weights from stratatlut
getwtvar.  If FALSE, strwtvar variable must be in stratalut.}

\item{getwtvar}{String. If getwt=TRUE, name of variable in stratalut to
calculate weights (Default = 'P1POINTCNT').}

\item{strwtvar}{String. If getwt=FALSE, name of variable in stratalut with
calculated weights (Default = 'strwt').}

\item{stratcombine}{Logical. If TRUE, and strata=TRUE, automatically combines
strata categories if less than minplotnum.strat plots in any one stratum. 
See notes for more info.}

\item{minplotnum.strat}{Integer. Minimum number of plots for a stratum
within an estimation unit.}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for strata.
}
\description{
Returns a list of user-supplied parameters and parameter values for strata.
}
\details{
If no parameters, an empty list is returned.
}
\examples{

strata_options(getwt = FALSE)

}
\author{
Grayson W. White
}
\keyword{options}
