% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{stunitco}
\alias{stunitco}
\title{SpatialPolygonsDataFrame with FIA state, unit, county codes and names}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3233 rows and 8 columns.
}
\source{
Downloaded from the United States Census Bureau on 2019 November 3,
format Esri Shapefile
(https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html)
Projection: Geographic (GCS_North_American_1983) EPSG: 4269
}
\usage{
stunitco
}
\description{
SpatialPolygonsDataFrame with FIA state, unit, county codes and names
}
\keyword{datasets}
