% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_options.R
\name{table_options}
\alias{table_options}
\title{Table aesthetics and output options.}
\usage{
table_options(
  row.FIAname = FALSE,
  col.FIAname = FALSE,
  row.orderby = NULL,
  col.orderby = NULL,
  row.add0 = FALSE,
  col.add0 = FALSE,
  rowlut = NULL,
  collut = NULL,
  rawdata = FALSE,
  rawonly = FALSE,
  rowgrp = FALSE,
  rowgrpnm = NULL,
  rowgrpord = NULL,
  sumunits = TRUE,
  allin1 = FALSE,
  metric = FALSE,
  estround = 1,
  pseround = 2,
  estnull = "--",
  psenull = "--",
  divideby = NULL,
  returntitle = FALSE,
  gainloss = FALSE,
  gainloss.vals = NULL,
  ...
)
}
\arguments{
\item{row.FIAname}{Logical. If TRUE, retrieves default FIA reference names
for rowvar located in FIESTA::ref_codes data frame. Names are only available
for certain variables (Check sort(unique(FIESTA::ref_codes$VARIABLE)) for
available names.  If row.FIAname = TRUE and rowvar is in FIESTA::ref_codes,
the rowvar name is used for the output table, and the rowvar code is used to
sort.}

\item{col.FIAname}{Logical. If TRUE, retrieves default FIA reference names
for colvar located in FIESTA::ref_codes data frame. Names are only available
for certain variables. Check: sort(unique(FIESTA::ref_codes$VARIABLE)) for
available names.  If col.FIAname = TRUE and rowvar is in FIESTA::ref_codes,
the colvar name is used for the output table, and the colvar code is used to
sort.}

\item{row.orderby}{String. Optional. Name of variable to sort table rows.
Both the rowvar and row.orderby variables must be included in the same input
data.frame.  if NULL, and row.FIAname=FALSE or rowvar is not in
FIESTA::ref_codes, the rows are ordered by rowvar.}

\item{col.orderby}{String. Optional. Name of variable to sort table columns.
Both the colvar and col.orderby variables must be included in the same input
data.frame.  if NULL, and col.FIAname=FALSE or colvar is not in
FIESTA::ref_codes, the columns are ordered by colvar.}

\item{row.add0}{Logical. If TRUE, include rows with 0 values to the output
table.}

\item{col.add0}{Logical. If TRUE, include columns with 0 values to the
output table.}

\item{rowlut}{Data frame. A lookup table with variable codes and code names
to include as rows of output table (See notes for more information and
format).}

\item{collut}{Data frame. A lookup table with variable codes and code names
to include as columns of output table (See notes for more information and
format).}

\item{rawdata}{Logical. If TRUE, returns a list of raw data tables that are
used for estimation (See Value). If savedata = TRUE, tables are written to
outfolder (if raw_fmt='csv') or raw_dsn (if raw_fmt != 'csv').}

\item{rawonly}{Logical. If TRUE, only rawdata are output. If dataset
includes many estimation units, and only raw data tables are desired, it is
more efficient to output raw data only.}

\item{rowgrp}{Logical. If TRUE, appends row groups to first column of table.
Only available if group category exists in ref_codes table or defined in
rowgrpnm (e.g., FORTYPGRPCD, OWNGRPCD).}

\item{rowgrpnm}{String. Name of variable for grouping rowvar. Variable must
be included in same input table as rowvar.}

\item{rowgrpord}{String. Name of variable to sort row group variable.
Variable must be included in same input table as rowgrpnm.}

\item{sumunits}{Logical. If TRUE, estimation units are summed and returned
in one table.}

\item{allin1}{Logical. If TRUE, both estimates and percent sample error are
output in one table as: estimates (percent sample error).}

\item{metric}{Logical. If TRUE, output area is in metric units (hectares).}

\item{estround}{Integer. Number of decimal places for estimates.}

\item{pseround}{Integer. Number of decimal places for percent sampling
error.}

\item{estnull}{Number or character. The number or symbol to use to indicate
'not sampled' for estimate.}

\item{psenull}{Number or character. The number or symbol to use to indicate
'not sampled' for percent standard errror.}

\item{divideby}{String. Conversion number for output ('hundred', 'thousand',
'million').}

\item{returntitle}{Logical. If TRUE, returns title(s) of the estimation
table(s).}

\item{gainloss}{Logical. If TRUE, a table with the difference of gain and
loss along with the variance and standard error, in percent, is generated.
Used in modGBlulc.}

\item{gainloss.vals}{String vector. A vector of names for values in gainloss
table. Used in modGBlulc.}

\item{...}{For extendibility.}

\item{title.main}{String. TITLE, if savedata=TRUE and/or returntitle=TRUE:
the complete title used for table. If title.main=NULL, the title.*
parameters are used to generate title string. Note: if title.ref is not
NULL, it is added to title.main.}

\item{title.ref}{String. TITLE, if savedata=TRUE and/or returntitle=TRUE:
the ending text of the table title (e.g., Nevada, 2004-2005). If NULL, = "".}

\item{title.rowvar}{String. TITLE, if savedata=TRUE and/or returntitle=TRUE:
pretty name for the row domain variable. If NULL, = rowvar.}

\item{title.colvar}{String. TITLE, if savedata=TRUE and/or returntitle=TRUE:
pretty name for the column domain variable. If NULL, = colvar.}

\item{title.unitvar}{String. TITLE, if savedata=TRUE and/or
returntitle=TRUE: pretty name for the estimation unit variable. If NULL, =
unitvar.}

\item{title.estvar}{String. TITLE: if savedata=TRUE and/or returntitle=TRUE:
pretty name for the estimate variable. If NULL, title.estvar = estvar.name.}

\item{title.filter}{String. TITLE, if savedata=TRUE and/or returntitle=TRUE:
pretty name for filter(s). If title.filter=NULL, a default is generated from
cfilter.  If title.filter="", no title.filter is used.}
}
\value{
A list of user-supplied parameters and parameter values for outputting
tables with custom aesthetics.
}
\description{
Returns a list of user-supplied parameters and parameter values for outputting
tables with custom aesthetics.
}
\details{
If no parameters, an empty list is returned.
}
\examples{

table_options(row.FIAnames = TRUE, col.FIAname = TRUE)

}
\author{
Grayson W. White
}
\keyword{options}
